backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Firebase Admin SDK
firebase-admin==6.4.0
google-cloud-firestore==2.13.1

# CrewAI (Optional - for production)
crewai==0.28.8

# Gemini AI (Optional - for questionnaire summarization)
google-generativeai==0.3.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client (optional)
httpx==0.25.1

# Date and time utilities
python-dateutil==2.8.2

